refactor: reorganize WebSocket endpoints into modular routers
- Extract WebSocket endpoints from monolithic webSocketEndpoints.ts into separate router modules
- Add 12 new router files organized by functionality:
  - baseRouter.ts: Base router class with common functionality
  - all.ts: Router aggregation and exports
  - encounter.ts: Combat and encounter management endpoints
  - entity.ts: Entity CRUD operations
  - fileSystem.ts: File system operations and management
  - macro.ts: Macro execution and management
  - pingPong.ts: Connection health check endpoints
  - roll.ts: Dice rolling and roll management
  - search.ts: Search functionality across game data
  - sheet.ts: Character sheet operations
  - structure.ts: Game structure and hierarchy management
  - utility.ts: Utility functions and helpers
- Update webSocketManager.ts to use new modular router structure
- Reduce webSocketEndpoints.ts from 2600+ lines to focused core functionality
- Improve code organization, maintainability, and separation of concerns
